legacy/requirements.txt
gunicorn
flask
pyyaml
psutil
tabulate
python-decouple==3.6